+++ lib/minglib/gui/edittext.hpp
@@ -5,10 +5,10 @@
 
 class EditText : public GUIObject {
 public:
-    EditText(long id, int x, int y, int width, int height);
-    EditText(long id, const char *text, int x, int y, int width, int height);
-    EditText(Window *parent, long id, int x, int y, int width, int height);
-    EditText(Window *parent, long id, const char *text, int x, int y, int width, int height);
+    EditText(long id, UINT16 x, UINT16 y, UINT16 width, UINT16 height);
+    EditText(long id, const char *text, UINT16 x, UINT16 y, UINT16 width, UINT16 height);
+    EditText(Window *parent, long id, UINT16 x, UINT16 y, UINT16 width, UINT16 height);
+    EditText(Window *parent, long id, const char *text, UINT16 x, UINT16 y, UINT16 width, UINT16 height);
 };
 
 #endif
